Release 2.2.9
Attempting this fix https://github.com/mbrn/material-table/issues/2164#issuecomment-658731507
